Refactoring the runner and copying nextflow.config
